Delete barcode.py && Create decetion_dmtx.py && Delete dijkstra.py && Update LineTracing.py && Create camera_test.py
