+++ detection_dmtx.py
@@ -1,0 +1,7 @@
+import cv2
+
+def Detection_dmtx(frame):
+    if(frame[max/2,max/2]):
+        return 1
+    else: 
+        return 0
